refactor: 전체 코드 변경
- 내부 styled-components를 styles.ts 파일로 분리
- 컴포넌트를 index.ts 파일에서 한번에 export하게끔 변경
