+++ src/Router.tsx
@@ -1,14 +1,12 @@
 import { Route, Routes } from "react-router-dom";
-import ChatRoute from "./routes/ChatRoute";
-import LoginRoute from "./routes/LoginRoute";
-import SignupRoute from "./routes/SignupRoute";
+import { ChatPage, LoginPage, SignupPage } from "./pages";
 
 const Router = () => {
   return (
     <Routes>
-      <Route path="/" element={<ChatRoute />} />
-      <Route path="/login" element={<LoginRoute />} />
-      <Route path="/signup" element={<SignupRoute />} />
+      <Route path="/" element={<ChatPage />} />
+      <Route path="/login" element={<LoginPage />} />
+      <Route path="/signup" element={<SignupPage />} />
     </Routes>
   );
 };
