+++ src/components/chat/MessageBox/index.ts
@@ -1,0 +1,2 @@
+export { default as MessageBox } from "./MessageBox";
+export { default as Message } from "./Message";
