+++ src/components/chat/index.ts
@@ -1,0 +1,1 @@
+export { default as Chat } from "./Chat";
