+++ src/components/chat/lobby/conversationList/index.ts
@@ -1,0 +1,2 @@
+export { default as ConversationList } from "./ConversationList";
+export { default as Conversation } from "./Conversation";
