+++ src/components/chat/lobby/index.ts
@@ -1,0 +1,1 @@
+export { default as Lobby } from "./Lobby";
