+++ src/components/chat/lobby/userList/index.ts
@@ -1,0 +1,2 @@
+export { default as UserList } from "./UserList";
+export { default as User } from "./User";
