+++ src/pages/SignupPage.tsx
@@ -1,7 +1,12 @@
-import Layout from "../components/signup/Layout";
+import { Signup } from "../components/signup";
+import St from "./styles";
 
 const SignupPage = () => {
-  return <Layout />;
+  return (
+    <St.PageContainer>
+      <Signup />
+    </St.PageContainer>
+  );
 };
 
 export default SignupPage;
