+++ src/pages/index.ts
@@ -1,0 +1,3 @@
+export { default as ChatPage } from "./ChatPage";
+export { default as LoginPage } from "./LoginPage";
+export { default as SignupPage } from "./SignupPage";
