--- src/routes/ChatRoute.tsx
@@ -1,7 +1,0 @@
-import ChatPage from "../pages/ChatPage";
-
-const ChatRoute = () => {
-  return <ChatPage />;
-};
-
-export default ChatRoute;
