--- src/routes/LoginRoute.tsx
@@ -1,7 +1,0 @@
-import LoginPage from "../pages/LoginPage";
-
-const LoginRoute = () => {
-  return <LoginPage />;
-};
-
-export default LoginRoute;
