--- src/routes/SignupRoute.tsx
@@ -1,7 +1,0 @@
-import SignupPage from "../pages/SignupPage";
-
-const SignupRoute = () => {
-  return <SignupPage />;
-};
-
-export default SignupRoute;
